I go hard in the mf paint
